[Refactor] StageDataManager 코드 리팩토링 롤백
