+++ JKC_Fallguys/Assets/1_Scripts/Scene/03_Stage/StageDataManager.cs
@@ -3,38 +3,80 @@ using UniRx;
 
 public class StageDataManager : SingletonMonoBehaviour<StageDataManager>
 {
-    public PlayerContainer PlayerContainer = new PlayerContainer(); 
-    
     // 게임 시작의 카운트다운을 관장합니다.
     private ReactiveProperty<bool> _isGameStart = new ReactiveProperty<bool>(false);
     public ReactiveProperty<bool> IsGameStart => _isGameStart;
 
     // 게임의 활성화 상태를 나타냅니다.
     private ReactiveProperty<bool> _isGameActive = new ReactiveProperty<bool>(false);
+
     public IReactiveProperty<bool> IsGameActive => _isGameActive;
+
+    // // 현재 클라이언트를 플레이하고 있는 플레이어의 생존 여부를 나타냅니다.
+    private Dictionary<int, ReactiveProperty<bool>> _isPlayerActiveDict = new Dictionary<int, ReactiveProperty<bool>>();
+
+    public IReactiveProperty<bool> IsPlayerActive(int actorNumber)
+    {
+        if (!_isPlayerActiveDict.ContainsKey(actorNumber))
+        {
+            _isPlayerActiveDict[actorNumber] = new ReactiveProperty<bool>(true);
+        }
+
+        return _isPlayerActiveDict[actorNumber];
+    }
 
     // 맵에서 쓰일 데이타가 저장되는 딕셔너리입니다. 
     public Dictionary<int, MapData> MapDatas = new Dictionary<int, MapData>();
-    
+
+    // 선택할 맵 데이터를 판별하는 객체입니다.
     public bool[] MapPickupFlags = new bool[DataManager.MaxPlayableMaps];
     private ReactiveProperty<int> _mapPickupIndex = new ReactiveProperty<int>();
+
     public IReactiveProperty<int> MapPickupIndex => _mapPickupIndex;
 
     // 플레이어의 점수들이 계속해서 저장되는 딕셔너리입니다.
     public Dictionary<int, PlayerData> PlayerDataByIndex = new Dictionary<int, PlayerData>();
+
     // 결과 창에서 사용될 플레이어의 인덱스를 캐싱해놓는 리스트입니다.
     public List<int> CachedPlayerIndicesForResults = new List<int>();
-    
+
     // 두 리스트는 스테이지가 넘어갈 때, 초기화됩니다.
     // 클리어에 실패한 사용자를 기록하는 리스트입니다
     public List<int> FailedClearStagePlayers = new List<int>();
+
     // 스테이지에서 사용될 순위를 기록하는 리스트입니다.
     public List<int> StagePlayerRankings = new List<int>();
+
+    // Round Result Panel의 성공, 실패, 종료 여부를 설정하기 위한 변수입니다.
+    public enum PlayerState
+    {
+        Default,
+        Victory,
+        Defeat,
+        GameTerminated
+    }
+
+    // 클라이언트 별 PlayerState를 관리하는 Dictionary입니다.
+    private Dictionary<int, ReactiveProperty<PlayerState>> _clientStates =
+        new Dictionary<int, ReactiveProperty<PlayerState>>();
+
+    public IReactiveProperty<PlayerState> GetCurrentState(int actorNumber)
+    {
+        if (!_clientStates.ContainsKey(actorNumber))
+        {
+            _clientStates[actorNumber] = new ReactiveProperty<PlayerState>();
+        }
+
+        return _clientStates[actorNumber];
+    }
+
+    // private ReactiveProperty<PlayerState> _currentState = new ReactiveProperty<PlayerState>();
+    // public IReactiveProperty<PlayerState> CurrentState => _currentState;
 
     // 라운드가 끝났는지 확인하기 위한 변수입니다.
     private ReactiveProperty<bool> _isRoundCompleted = new ReactiveProperty<bool>();
     public IReactiveProperty<bool> IsRoundCompleted => _isRoundCompleted;
-    
+
     public void AddPlayerToRanking(int playerIndex)
     {
         StagePlayerRankings.Add(playerIndex);
@@ -63,6 +105,16 @@ public class StageDataManager : SingletonMonoBehaviour<StageDataManager>
         _isRoundCompleted.Value = status;
     }
 
+    public void SetPlayerState(int actorNumber, PlayerState state)
+    {
+        GetCurrentState(actorNumber).Value = state;
+    }
+
+    public void SetPlayerActive(int actorNumber, bool status)
+    {
+        IsPlayerActive(actorNumber).Value = status;
+    }
+
     /// <summary>
     /// StateDataManager는 Singleton으로 구성되어 있습니다.
     /// 이 클래스는, Loading이 시작될 때 생성되고 Lobby로 돌아갈 경우 파괴되어야 합니다.
@@ -83,7 +135,7 @@ public class StageDataManager : SingletonMonoBehaviour<StageDataManager>
                 ++index;
             }
         }
-        
+
         return index == 3;
     }
     
@@ -91,7 +143,7 @@ public class StageDataManager : SingletonMonoBehaviour<StageDataManager>
     {
         _mapPickupIndex.Value = index;
     }
-
+    
     public void SetGameStart(bool status)
     {
         _isGameStart.Value = status;
