+++ JKC_Fallguys/Assets/Resources/Prefabs/Stage/FruitChute/MapFruitChute.prefab
@@ -5834,7 +5834,6 @@ GameObject:
   m_Component:
   - component: {fileID: 3496625069096609031}
   - component: {fileID: 717357563782364298}
-  - component: {fileID: 5377920278310531889}
   - component: {fileID: 5102425667953564387}
   m_Layer: 0
   m_Name: Camera
@@ -5901,14 +5900,6 @@ Camera:
   m_OcclusionCulling: 1
   m_StereoConvergence: 10
   m_StereoSeparation: 0.022
---- !u!81 &5377920278310531889
-AudioListener:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6500085074855322018}
-  m_Enabled: 1
 --- !u!114 &5102425667953564387
 MonoBehaviour:
   m_ObjectHideFlags: 0
